% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data-structures.R
\name{convert_names}
\alias{convert_names}
\title{Convert names}
\usage{
convert_names(x)
}
\arguments{
\item{x}{A vector of names}
}
\description{
Convert names for compatibility in R, removing promblematic characters such as "+" of spaces
}

