% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/data-structures.R
\name{display_parameters}
\alias{display_parameters}
\title{Display populations and reagents
Provides a list of populations and reagents that can be used to select which are desired features
for subsequent analysis}
\usage{
display_parameters(FCS_files, experimentID, access_key)
}
\arguments{
\item{FCS_files}{Object containing FCS files}

\item{experimentID}{the experiment ID as a string}

\item{access_key}{An object containing server URL and authentication info: see "authenticate"}
}
\description{
Display populations and reagents
Provides a list of populations and reagents that can be used to select which are desired features
for subsequent analysis
}

