% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/base-API-functions.R
\name{downloadFcsFile}
\alias{downloadFcsFile}
\title{Download an FCS file}
\usage{
downloadFcsFile(experimentId, fcsFileId, localFilePath)
}
\arguments{
\item{experimentId}{the experiment ID as a string}

\item{fcsFileId}{the FCS file ID as a string}

\item{localFilePath}{path where to download the data}
}
\description{
Download an FCS file
}

