% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data-structures.R
\name{get_annotations}
\alias{get_annotations}
\title{Get annotations for a set of FCS files}
\usage{
get_annotations(FCS_files)
}
\arguments{
\item{FCS_files}{object containing FCS files from get_FCS_files()}
}
\description{
If some files are not annotated, it will return a warning and only return the fcs files and annotations only
}

