% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/base-API-functions.R
\name{get_events}
\alias{get_events}
\title{Retrieves events}
\usage{
get_events(experimentID, FCS_fileID, access_key, populationID = NULL)
}
\arguments{
\item{experimentID}{the experiment ID as a string}

\item{FCS_fileID}{the FCS file ID as a string}

\item{access_key}{An object containing server URL and authentication info: see "authenticate"}

\item{populationID}{the population ID as a string}
}
\description{
Retrieves events
}

