% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/base-API-functions.R
\name{get_experiment}
\alias{get_experiment}
\title{Retrieve one experiment}
\usage{
get_experiment(experimentID, access_key)
}
\arguments{
\item{experimentID}{the experiment ID as a string}

\item{access_key}{An object containing server URL and authentication info: see "authenticate"}
}
\description{
This function retrieves an object with the information from one experiment specified by the experiment ID
}

