% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/base-API-functions.R
\name{get_experiments}
\alias{get_experiments}
\title{Retrieve all accessible experiments}
\usage{
get_experiments(access_key)
}
\arguments{
\item{access_key}{an object containing server URL and authentication info: see "authenticate"}

\item{baseURL}{URL for the server}
}
\description{
This function retrieves an object with the information about all acessible experiments
}
\examples{
access_key <- authenticate("my_username", "my_password",
                           baseURL =  "http://52.27.144.218/api/v1")
experiments <- get_experiments(access_key)
}

