% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/base-API-functions.R
\name{get_FCS_files}
\alias{get_FCS_files}
\title{Get all FCS files for an experiment}
\usage{
get_FCS_files(experimentID, access_key)
}
\arguments{
\item{experimentID}{the experiment ID as a string}

\item{access_key}{An object containing server URL and authentication info: see "authenticate"}
}
\description{
Returns an object containing all FCS files from the experiments based on the experiment ID
}

