% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data-structures.R
\name{get_folds}
\alias{get_folds}
\title{Generate fold change statistics object}
\usage{
get_folds(statistics_object, basal_name, fold_type, unique_IDs = c("file_ID",
  "filename"), annotation_columns = c("Donor", "Species", "Gender",
  "Condition"))
}
\arguments{
\item{statistics_object}{data frame of statistics}

\item{basal_name}{name of the basal condition as string e.g. "Basal"}

\item{fold_type}{either "asinh", calculates asinh ratio, or "raw", fold of raw counts}

\item{unique_IDs}{a vector of strings that are the names of the unique identifier columns}

\item{annotation_columns}{vector of strings that are the names of the annotation columns}
}
\description{
Takes a statistics object and presents statistics relative to designated baseline.
Options exist for asinh ratio or fold change on raw counts.
If using a statistics object that is an output of a get_statistics function, use statistics_object$statistics
Must contain "Condition" column
Unique identifier columns such as filename, File ID, etc must be specified as arguments for removal
}

