% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/base-API-functions.R
\name{get_plot}
\alias{get_plot}
\title{Plot the data}
\usage{
get_plot(experimentID, FCS_fileID, access_key, x_channel_name, y_channel_name,
  scale_setID, populationID = NULL, plot_type = "density", ticks = "true",
  tick_labels = "false", axis_labels = "true")
}
\arguments{
\item{experimentID}{the experiment ID as a string}

\item{FCS_fileID}{the FCS file ID as a string}

\item{access_key}{An object containing server URL and authentication info: see "authenticate"}

\item{x_channel_name}{The channel for the x axis named by metal}

\item{y_channel_name}{The channel for the y axis named by metal}

\item{populationID}{the population ID as a string}
}
\description{
Creates a plot of specified channels and populations
}

