% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/base-API-functions.R
\name{get_populations}
\alias{get_populations}
\title{Get all populations for an experiment}
\usage{
get_populations(experimentID, acess_key)
}
\arguments{
\item{experimentID}{the experiment ID as a string}

\item{access_key}{An object containing server URL and authentication info: see "authenticate"}
}
\description{
Get all populations for an experiment
}

