% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/data-structures.R
\name{get_statistics_set}
\alias{get_statistics_set}
\title{Get set of statistics for chosen features}
\usage{
get_statistics_set(experimentID, FCS_files, access_key, populations, reagents,
  statistic_type, k = NULL, annotate = FALSE)
}
\arguments{
\item{experimentID}{the experiment ID as a string}

\item{access_key}{An object containing server URL and authentication info: see "authenticate"}

\item{populations}{A character vector of population names}

\item{reagents}{A character vector of reagent names}

\item{statistic_type}{Statistic desired, "mean", "median", "quantile", "eventcount"}

\item{k}{required for statistic "quantile", number from 0.0 to 1.0}

\item{annotate}{Logical operator to indicate whether annotations should be included in returned object}

\item{FCS_fileID}{the FCS file ID as a string}
}
\description{
Specify populations and reagents to generate a feature set of each pair and retrieve statistics
Annotations are optional
}

