% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data-structures.R
\name{get_statistics_set_parallel}
\alias{get_statistics_set_parallel}
\title{Get set of statistics for chosen features}
\usage{
get_statistics_set_parallel(experimentID, FCS_files, access_key, populations,
  reagents, statistic_type, k = NULL, b = batch_size, annotate = FALSE)
}
\arguments{
\item{experimentID}{the experiment ID as a string}

\item{access_key}{an object containing server URL and authentication info: see "authenticate"}

\item{populations}{character vector of population names}

\item{reagents}{character vector of reagent names}

\item{statistic_type}{statistic desired, "mean", "median", "quantile", "eventcount"}

\item{k}{required for statistic "quantile", number from 0.0 to 1.0}

\item{annotate}{logical operator to indicate whether annotation00s should be included in returned object}

\item{FCS_fileID}{the FCS file ID as a string}
}
\description{
Specify populations and reagents to generate a feature set of each pair and retrieve statistics
Annotations are optional
}
\examples{
access_key <- authenticate("my_username", "my_password",
                           baseURL =  "http://52.27.144.218/api/v1")
experiments <- get_experiments(access_key)
experimentID <- experiments[experiments$name == "experiment_name", '_id']
FCS_fileID <- get_FCS_files(experimentID, access_key)[1, "_id"]
populations <- c("CD4+T cells", "CD8+T cells", "CD14+ Monocytes")
reagents <- c("pSTAT1", "pSTAT3", "pMAPKAPK2")
get_statistic_set(experimentID, FCS_files, access_key, populations, reagents,
                   statistic_type = "median", k = NULL, annotate = TRUE)
}

