% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/base-API-functions.R
\name{getEvents}
\alias{getEvents}
\title{Retrieves events}
\usage{
getEvents(experimentId, fcsFileId, populationId = NULL)
}
\arguments{
\item{experimentId}{the experiment ID as a string}

\item{fcsFileId}{the FCS file ID as a string}

\item{populationId}{the population ID as a string}
}
\description{
Retrieves events
}

