% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/base-API-functions.R
\name{getExperiment}
\alias{getExperiment}
\title{Retrieve one experiment}
\usage{
getExperiment(experimentId)
}
\arguments{
\item{experimentId}{the experiment ID as a string}
}
\description{
This function retrieves an object with the information from one experiment specified by the experiment ID
}

