% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/base-API-functions.R
\name{getExperiments}
\alias{getExperiments}
\title{Retrieve all accessible experiments}
\usage{
getExperiments()
}
\description{
This function retrieves an object with the information about all acessible experiments
}

