% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/base-API-functions.R
\name{getFcsFilesForExperiment}
\alias{getFcsFilesForExperiment}
\title{Get all FCS files for an experiment}
\usage{
getFcsFilesForExperiment(experimentId)
}
\arguments{
\item{experimentId}{the experiment ID as a string}
}
\description{
Returns an object containing all FCS files from the experiments based on the experiment ID
}

