% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/base-API-functions.R
\name{getGatesForExperiment}
\alias{getGatesForExperiment}
\title{Get all gates for an experiment}
\usage{
getGatesForExperiment(experimentId)
}
\arguments{
\item{experimentId}{the experiment ID as a string}
}
\description{
Get all gates for an experiment
}

