% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/base-API-functions.R
\name{getPlot}
\alias{getPlot}
\title{Plot the data}
\usage{
getPlot(experimentId, fcsFileId, xChannelName, yChannelName, scaleSetId,
  populationId = NULL, plotType = "density", ticks = "true",
  tickLabels = "false", axisLabels = "true")
}
\arguments{
\item{experimentId}{the experiment ID as a string}

\item{fcsFileId}{the FCS file ID as a string}

\item{xChannelName}{The channel for the x axis named by metal}

\item{yChannelName}{The channel for the y axis named by metal}

\item{populationId}{the population ID as a string}
}
\description{
Creates a plot of specified channels and populations
}

