% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/base-API-functions.R
\name{getPopulationsForExperiment}
\alias{getPopulationsForExperiment}
\title{Get all populations for an experiment}
\usage{
getPopulationsForExperiment(experimentId)
}
\arguments{
\item{experimentId}{the experiment ID as a string}
}
\description{
Get all populations for an experiment
}

