% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/base-API-functions.R
\name{getScaleSetsForExperiment}
\alias{getScaleSetsForExperiment}
\title{Get all scale sets for an experiment}
\usage{
getScaleSetsForExperiment(experimentId)
}
\arguments{
\item{experimentId}{the experiment ID as a string}
}
\description{
Get all scale sets for an experiment
}

