% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/base-API-functions.R
\name{getStatistic}
\alias{getStatistic}
\title{Retrieve statistic from a population}
\usage{
getStatistic(experimentId, fcsFileId, channelName, statisticType, k = NULL,
  populationId = NULL)
}
\arguments{
\item{experimentId}{the experiment ID as a string}

\item{fcsFileId}{the FCS file ID as a string}

\item{channelName}{Channel name}

\item{statisticType}{Statistic desired, "mean", "median", "quantile", "eventcount}

\item{k}{required for statistic "quantile", number from 0.0 to 1.0}
}
\description{
Returns a desired statistic from a desired feature
}

