% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/base-API-functions.R
\name{authenticate}
\alias{authenticate}
\title{Authenticate acess}
\usage{
authenticate(username, password, baseURL)
}
\arguments{
\item{username}{username for the account}

\item{password}{password for the account}

\item{baseURL}{URL for the server}
}
\description{
This function establishes the connection to your account on the remote server
It returns an object containing the server URL and your connection information
}
\examples{
authenticate("my_username", "my_password",
             baseURL =  "http://52.27.144.218/api/v1")
}

