% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/base-API-functions.R
\name{authenticate}
\alias{authenticate}
\title{Authenticate acess}
\usage{
authenticate(username, password)
}
\arguments{
\item{username}{The username for the account}

\item{password}{The password for the account}
}
\description{
This function establishes the connection to your account on the remote server
}

