% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data-structures.R
\name{display_parameters}
\alias{display_parameters}
\title{Display populations and reagents}
\usage{
display_parameters(FCS_files, experimentID, access_key)
}
\arguments{
\item{FCS_files}{Object containing FCS files}

\item{experimentID}{the experiment ID as a string}

\item{access_key}{An object containing server URL and authentication info: see "authenticate"}
}
\description{
Provides a list of population names, channel names, and reagent names that can be used to select which are desired features
for subsequent analysis
}

