% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/base-API-functions.R
\name{download_fcs_file}
\alias{download_fcs_file}
\title{Download an FCS file}
\usage{
download_fcs_file(experimentID, FCS_fileID, local_file_path, access_key)
}
\arguments{
\item{experimentID}{the experiment ID as a string}

\item{FCS_fileID}{the FCS file ID as a string}

\item{local_file_path}{path where to download the data including file name}

\item{access_key}{an object containing server URL and authentication info: see "authenticate"}
}
\description{
Download an FCS file
}

