% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data-structures.R
\name{get_annotations}
\alias{get_annotations}
\title{Get annotations for a set of FCS files}
\usage{
get_annotations(FCS_files)
}
\arguments{
\item{FCS_files}{Object containing FCS files}
}
\description{
Get annotations for a set of FCS files
}

