% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/base-API-functions.R
\name{get_experiment}
\alias{get_experiment}
\title{Retrieve one experiment}
\usage{
get_experiment(experimentID, access_key)
}
\arguments{
\item{experimentID}{the experiment ID as a string}

\item{access_key}{an object containing server URL and authentication info: see "authenticate"}
}
\description{
This function retrieves an object with the information from one experiment specified by the experiment ID
}
\examples{
access_key <- authenticate("my_username", "my_password",
                           baseURL =  "http://52.27.144.218/api/v1")
experiments <- get_experiments(access_key)
experimentID <- experiments[experiments$name == "experiment_name", '_id']
get_experiment(experimentID, access_key)
}

