% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/base-API-functions.R
\name{get_experiments}
\alias{get_experiments}
\title{Retrieve all accessible experiments}
\usage{
get_experiments(access_key)
}
\arguments{
\item{access_key}{An object containing server URL and authentication info: see "authenticate"}

\item{baseURL}{The URL for the server}
}
\description{
This function retrieves an object with the information about all acessible experiments
}

