% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/base-API-functions.R
\name{get_FCS_files}
\alias{get_FCS_files}
\title{Get all FCS files for an experiment}
\usage{
get_FCS_files(experimentID, access_key)
}
\arguments{
\item{experimentID}{the experiment ID as a string}

\item{access_key}{an object containing server URL and authentication info: see "authenticate"}
}
\description{
Returns a data frame containing all FCS files from an experiment based on the experiment ID
Lists information about event count, panel, and annotations
}
\examples{
access_key <- authenticate("my_username", "my_password",
                           baseURL =  "http://54.186.1.226/api/v1")
experiments <- get_experiments(access_key)
experimentID <- experiments[experiments$name == "experiment_name", '_id']
get_FCS_files(experimentID, access_key)
}

