% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data-structures.R
\name{get_folds}
\alias{get_folds}
\title{Generate fold change statistics object}
\usage{
get_folds(statistics_object, basal_name, fold_type)
}
\arguments{
\item{statistics_object}{object output of get_statistics function}

\item{basal_name}{name of the basal condition as string e.g. "Basal"}

\item{fold_type}{either "asinh", calculates asinh ratio, or "raw", fold of raw counts}
}
\description{
Takes a statistics object and presents statistics relative to designated baseline
Options exist for asinh ratio or fold change on raw counts
}

