% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/base-API-functions.R
\name{get_gates}
\alias{get_gates}
\title{Get all gates for an experiment}
\usage{
get_gates(experimentID, access_key)
}
\arguments{
\item{experimentID}{the experiment ID as a string}

\item{access_key}{an object containing server URL and authentication info: see "authenticate"}
}
\description{
Returns a data frame with the information on all gates from an experiment including names, dimensions, tailored, etc.
}

