% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/base-API-functions.R
\name{get_populations}
\alias{get_populations}
\title{Get all populations for an experiment}
\usage{
get_populations(experimentID, acess_key)
}
\arguments{
\item{experimentID}{the experiment ID as a string}

\item{access_key}{an object containing server URL and authentication info: see "authenticate"}
}
\description{
Get all populations for an experiment
}
\examples{
access_key <- authenticate("my_username", "my_password",
                           baseURL =  "http://52.27.144.218/api/v1")
experiments <- get_experiments(access_key)
experimentID <- experiments[experiments$name == "experiment_name", '_id']
get_populations(experimentID, access_key)
}

