% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/base-API-functions.R
\name{get_scale_sets}
\alias{get_scale_sets}
\title{Get all scale sets for an experiment}
\usage{
get_scale_sets(experimentID, access_key)
}
\arguments{
\item{experimentID}{the experiment ID as a string}

\item{access_key}{An object containing server URL and authentication info: see "authenticate"}
}
\description{
Get all scale sets for an experiment
}

