% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/base-API-functions.R
\name{get_statistic}
\alias{get_statistic}
\title{Retrieve statistic from a population}
\usage{
get_statistic(experimentID, FCS_fileID, access_key, channel_name,
  statistic_type, k = NULL, populationID = NULL)
}
\arguments{
\item{experimentID}{the experiment ID as a string}

\item{FCS_fileID}{the FCS file ID as a string}

\item{access_key}{An object containing server URL and authentication info: see "authenticate"}

\item{channel_name}{Channel name}

\item{statistic_type}{Statistic desired, "mean", "median", "quantile", "eventcount}

\item{k}{required for statistic "quantile", number from 0.0 to 1.0}

\item{populationID}{Population ID of population of interest}
}
\description{
Returns a desired statistic from a desired feature
}
\examples{
access_key <- authenticate("my_username", "my_password",
                           baseURL =  "http://52.27.144.218/api/v1")
experiments <- get_experiments(access_key)
experimentID <- experiments[experiments$name == "experiment_name", '_id']
FCS_fileID <- get_FCS_files(experimentID, access_key)[1, "_id"]
get_statistic(experimentID, FCS_fileID, access_key, channel_name = "La139i",
               statistic_type = "median", k=NULL, populationID=NULL)
}

